requirementrs.txt
pandas
SQLAlchemy
python-dotenv
pandas-ta
setuptools